random finished.

+++ component.h
@@ -7,8 +7,8 @@ class component
 private:
 	typeComp _typeComp;
 protected:
-	unsigned int workTimeSec;//Работа компонента в секундах
-	float var;//Стоимость компонента
+	int workTimeSec;//Работа компонента в секундах
+	int var;//Стоимость компонента
 public:
 	typeComp Set_rand_type();
 	component();
